Separate transformation tests (#488)
